fix(kernel): real witness points, principal directions, surface centroid in occt-wasm (#809)
## Summary

Three measurement queries on the `occt-wasm` adapter previously returned
placeholder values because the C++ facade exposes only partial
information. This PR replaces all three placeholders with real
computations done in TS, using primitives the facade already exposes
(tessellation, point-on-surface, vertex positions).

- **`distance()`**: `point1`/`point2` were `[0, 0, 0]`. Now sampled from
topological vertices + tessellation; the closest pair is returned as
approximate witness points. The scalar `value` stays exact
(`BRepExtrema` in C++).
- **`surfaceCurvature()`**: `maxDirection`/`minDirection` were `[1, 0,
0]`/`[0, 1, 0]`. Now derived from the shape operator `W = II · I⁻¹` in
the `{Pu, Pv}` basis, with finite-difference fundamental forms.
Isotropic and degenerate-parametrisation cases handled.
- **`surfaceCenterOfMass()`**: averaged vertex positions. Now an
area-weighted triangle centroid via `tessellate()` — exact for the
triangulated face.

All three are pure-TS — the external `occt-wasm` npm package's C++
facade is unchanged. Smooth analytic surfaces (sphere, cylinder) match
within ~1% at default tessellation deflection.

## Test plan

- [x] `npm run typecheck`
- [x] `npm run check:patterns`
- [x] `npm run check:boundaries`
- [x] `npm run format:check`
- [x] `npx vitest run tests/measureFns.test.ts` — passes on all 3 kernel
projects (`occt`, `brepkit`, `occt-wasm`)
- [x] Full suite: 0 regressions vs. baseline; one previously-failing
`occt-wasm` test now passes
- [x] New tests verify analytic ground truth:
  - cylinder lateral centroid on the cylinder axis
- sphere face centroid at the sphere center (skipped on brepkit due to a
separate `tessellateFace` seam-asymmetry issue)
  - sphere principal curvatures = ±1/R, unit-length tangent directions
  - planar face has zero principal curvatures
  - cylinder lateral face has one curvature ≈ 1/R, the other ≈ 0
- box-pair witness points lie on the connecting line, witness distance ≥
exact distance

+++ src/kernel/occt/meshOps.ts
@@ -25,7 +25,7 @@ import { perfTimer } from '../perfStats.js';
 function sliceF32(heap: Float32Array, ptr: number, size: number): Float32Array {
   if (size === 0) return new Float32Array(0);
   const offset = ptr / 4;
-  return heap.slice(offset, offset + size) as Float32Array;
+  return heap.slice(offset, offset + size);
 }
 
 /**
